adjust DefinePlugins and configs for different use cases (debugging device drivers vs. UI)
