+++ config/device-debug.sh
@@ -1,6 +1,6 @@
 export API_URL='http://localhost:8009'
 export UPLOAD_URL='http://localhost:9122'
 export BLIP_URL='http://localhost:3000'
-export DEBUG_ERROR=false
+export DEBUG_ERROR=true
 export REDUX_LOG=false
 export REDUX_DEV_UI=false
