+++ config/ui-debug.sh
@@ -1,4 +1,6 @@
 export API_URL='http://localhost:8009'
 export UPLOAD_URL='http://localhost:9122'
 export BLIP_URL='http://localhost:3000'
-export DEBUG_ERROR=true
+export DEBUG_ERROR=false
+export REDUX_LOG=true
+export REDUX_DEV_UI=true
