+++ lib/containers/root/Root.js
@@ -15,9 +15,9 @@
  * == BSD2 LICENSE ==
  */
 
-/* global __DEBUG__ */
+/* global __REDUX_DEV_UI__ */
 
-if (__DEBUG__ === true) {
+if (__REDUX_DEV_UI__ === true) {
   module.exports = require('./Root.dev');
 } else {
   module.exports = require('./Root.prod');
