+++ lib/redux/store/configureStore.js
@@ -15,10 +15,43 @@
  * == BSD2 LICENSE ==
  */
 
-/* global __DEBUG__ */
+/* global __REDUX_DEV_UI__, __REDUX_LOG__ */
 
-if (__DEBUG__ === true) {
-  module.exports = require('./configureStore.dev');
+import { compose, createStore } from 'redux';
+
+import api from '../../core/api';
+import config from '../../config';
+api.create({
+  apiUrl: config.API_URL,
+  uploadUrl: config.UPLOAD_URL,
+  version: config.version
+});
+
+import uploader from '../reducers/';
+
+import DevTools from '../../components/DevTools';
+
+let middlewares;
+
+if (__REDUX_LOG__ === true) {
+  middlewares = require('./middlewares.dev.js')(api);
 } else {
-  module.exports = require('./configureStore.prod');
+  middlewares = require('./middlewares.prod.js')(api);
+}
+
+let finalCreateStore;
+
+if (__REDUX_DEV_UI__ === true) {
+  finalCreateStore = compose(
+    middlewares,
+    DevTools.instrument()
+  )(createStore);
+}
+else {
+  finalCreateStore = compose(middlewares)(createStore);
+}
+
+export default function configureStore(initialState) {
+  const store = finalCreateStore(uploader, initialState);
+  return { api, store };
 }
