+++ lib/redux/store/middlewares.dev.js
@@ -15,42 +15,23 @@
  * == BSD2 LICENSE ==
  */
 
-import { applyMiddleware, compose, createStore } from 'redux';
+
+import { applyMiddleware } from 'redux';
 import createLogger from 'redux-logger';
 import thunk from 'redux-thunk';
 
 import { createErrorLogger } from '../utils/errors';
 import { createMetricsTracker } from '../utils/metrics';
 
-import api from '../../core/api';
-import config from '../../config';
-api.create({
-  apiUrl: config.API_URL,
-  uploadUrl: config.UPLOAD_URL,
-  version: config.version
-});
-
-import uploader from '../reducers/';
-
-import DevTools from '../../components/DevTools';
-
-const finalCreateStore = compose(
-  /*
-   * order is significant here!
-   * in particular, the thunk middleware must be applied first
-   * redux middleware doc is a work of easily-understood genius:
-   * http://redux.js.org/docs/advanced/Middleware.html
-   */
-  applyMiddleware(
+export default (api) => applyMiddleware(
+    /*
+     * order is significant here!
+     * in particular, the thunk middleware must be applied first
+     * redux middleware doc is a work of easily-understood genius:
+     * http://redux.js.org/docs/advanced/Middleware.html
+     */
     thunk,
     createLogger(),
     createErrorLogger(api),
     createMetricsTracker(api)
-  ),
-  DevTools.instrument()
-)(createStore);
-
-export default function configureStore(initialState) {
-  const store = finalCreateStore(uploader, initialState);
-  return { api, store };
-}
+);
