+++ lib/redux/store/middlewares.prod.js
@@ -15,37 +15,21 @@
  * == BSD2 LICENSE ==
  */
 
-import { applyMiddleware, compose, createStore } from 'redux';
+
+import { applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 
 import { createErrorLogger } from '../utils/errors';
 import { createMetricsTracker } from '../utils/metrics';
 
-import api from '../../core/api';
-import config from '../../config';
-api.create({
-  apiUrl: config.API_URL,
-  uploadUrl: config.UPLOAD_URL,
-  version: config.version
-});
-
-import uploader from '../reducers';
-
-const finalCreateStore = compose(
-  /*
-   * order is significant here!
-   * in particular, the thunk middleware must be applied first
-   * redux middleware doc is a work of easily-understood genius:
-   * http://redux.js.org/docs/advanced/Middleware.html
-   */
-  applyMiddleware(
+export default (api) => applyMiddleware(
+    /*
+     * order is significant here!
+     * in particular, the thunk middleware must be applied first
+     * redux middleware doc is a work of easily-understood genius:
+     * http://redux.js.org/docs/advanced/Middleware.html
+     */
     thunk,
     createErrorLogger(api),
     createMetricsTracker(api)
-  )
-)(createStore);
-
-export default function configureStore(initialState) {
-  const store = finalCreateStore(uploader, initialState);
-  return { api, store };
-}
+);
