+++ webpack.config.js
@@ -4,6 +4,8 @@ var webpack = require('webpack');
 
 var definePlugin = new webpack.DefinePlugin({
   __DEBUG__: JSON.stringify(JSON.parse(process.env.DEBUG_ERROR || 'false')),
+  __REDUX_LOG__: JSON.stringify(JSON.parse(process.env.REDUX_LOG || 'false')),
+  __REDUX_DEV_UI__: JSON.stringify(JSON.parse(process.env.REDUX_DEV_UI || 'false')),
   __TEST__: false
 });
